requirements.txt
# Core dependencies
streamlit>=1.28.0
PyPDF2>=3.0.0

# LangChain ecosystem
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Vector store and embeddings
faiss-cpu>=1.7.4

# OpenAI client
openai>=1.0.0

# Environment management
python-dotenv>=1.0.0
